Add P2 cleanup: symlink extraction, interchange levels, multi-extent reads
Extract now materializes Rock Ridge symbolic links as real symlinks
instead of empty files, replacing stale targets from earlier runs. The
symlink target travels on FileSystemEntry, populated by both the parser
and the tree-backed entry refresh.

Add ISO 9660 interchange level support: WithInterchangeLevel(1|2|3)
selects the identifier rules enforced at pack time. Level 1 switches the
Rock Ridge mangler to 8.3 output; without Rock Ridge, invalid raw
identifiers are rejected with a clear error instead of writing a
non-conforming image. Levels 2 and 3 enforce the 31-character budget.
Validation lives in the validation package (ValidateFileIdentifier).

Add multi-extent file assembly on the read path: consecutive directory
records flagged MultiExtent merge into a single entry backed by a
segmented reader that presents the concatenated extents as one stream.
Entry and tree node sizes are promoted to uint64 accordingly; writing a
file larger than a single 32-bit extent now fails with an explicit
multi-extent-write-not-implemented error rather than corrupting sizes.
Also fixes a uint16 overflow lurking in boot entry size reporting.

Implement VolumePartitionDescriptor Marshal/Unmarshal per ECMA-119 8.6
(previously stubbed to empty) and add VolumeDescriptorSet.Validate for
structural invariant checks.

Tests: symlink extraction to disk, 8.3 mangling, strict-level rejection,
segmented reader boundary/EOF behavior, parser multi-extent merging via
a crafted directory extent, VPD byte-exact round-trip, and descriptor
set validation.

+++ pkg/iso9660/boot/eltorito.go
@@ -467,7 +467,7 @@ func (et *ElTorito) BuildBootImageEntries() ([]*filesystem.FileSystemEntry, erro
 			Name:       filename,
 			FullPath:   "/[BOOT]/" + filename, // Logical path inside the ISO
 			IsDir:      false,
-			Size:       uint32(entry.size) * 512, // Convert 512-byte block size
+			Size:       uint64(entry.size) * 512, // Convert 512-byte block size
 			Location:   entry.location,
 			Mode:       0444,        // Read-only boot image
 			CreateTime: time.Time{}, // No real timestamp in El Torito
